[ELY-136] Move ByteStringBuilder to the common util package.
