+++ src/test/java/org/wildfly/security/sasl/test/ByteStringBuilderTest.java
@@ -26,7 +26,7 @@ import javax.crypto.spec.SecretKeySpec;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.wildfly.security.sasl.util.ByteStringBuilder;
+import org.wildfly.security.util.ByteStringBuilder;
 
 /**
  * Tests of org.wildfly.security.sasl.util.ByteStringBuilder
